rao/requirements.txt
django==2.2.23
django-extensions==2.2.9
reportlab==3.5.42
python-codicefiscale==0.3.7
xhtml2pdf==0.2.4
Werkzeug==1.0.1
jwcrypto==0.7
PyJWT==1.7.1
pyOpenSSL==19.1.0
python3-saml==1.9.0
requests==2.24.0
cryptography==2.9.2
djangorestframework==3.12.4
